Fix packaging issue — Windows
Maven cannot execute the shell script on Windows because
it is not an executable file in the Windows sense
→ `.exe`, `.com`, `.cmd`, `.bat`

I fix this by executing directly the Bash interpreter.

NB:

I don't use any Batch or Powershell script to compile files or execute scripts.
If you work on Windows, you must to use a portable console emulator (Cmder, or Git Bash).

Link:

- https://stackoverflow.com/questions/25511402/java-io-ioexception-cannot-run-program-createprocess-error-193-1-is-not-a