Change the architecture of the program
The Evaluator is becoming a god object.
In order to prevent a violation of the Single Responsibility principle (SRP),
I change the architeture of the program.

> Consequences

- Evaluator is responsible of evaluating an expression
- A new class is created : Normalizer
- For each operation, a related class will be constructed
    - substitution (WIP as of this commit)
    - derivative (not implemented yet)
    - primitive (not implemented yet)
    - ...

> The normalizer class

The normalizer integrates some elements from the evaluator, i.e:

- the environment
- The help menu

It is responsible of "normalizing" the Abstract Syntax Tree (AST), that is to say
executing a command or eveluate an expression.

The evaluator does not need to manage the environment anymore, it will use it to retrieve variables.