Create a new sealed class
Every math functions has an expression in argument in common.
This common property is not used properly by the program.
The consequence of that is a code duplication in some classes.
This is particularly true in the `when` pattern matching.

That is why the FunCall object is created so it can store an abstract property
shared by every subclasses.